Fix: Correct types for compatibility with 64bit windows.
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk@555 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/CDlgCancel.cpp
@@ -23,7 +23,7 @@ CDlgCancel::CDlgCancel()
 int CDlgCancel::DoModal( HINSTANCE hInstance, HWND hwndParent, int nDlgTemplete )
 {
 	m_bCANCEL = FALSE;	/* IDCANCELボタンが押された */
-	return CDialog::DoModal( hInstance, hwndParent, nDlgTemplete, NULL );
+	return (int)CDialog::DoModal( hInstance, hwndParent, nDlgTemplete, NULL );
 }
 /* モードレスダイアログの表示 */
 HWND CDlgCancel::DoModeless( HINSTANCE hInstance, HWND hwndParent, int nDlgTemplete )
